plane avatar class - to separate module
